+++ src/api-legacy/v1/borehole/producer.py
@@ -191,9 +191,8 @@ class BoreholeProducerHandler(Producer):
                         request['orderby'] is not None
                     ) and (
                         'orderby' in self.user['setting']['eboreholetable'] and
-                        self.user[
-                            'setting'
-                        ]['eboreholetable']['orderby'] != request['orderby']
+                        ('orderby' not in self.user['setting']['eboreholetable'] or
+                        self.user['setting']['eboreholetable']['orderby'] != request['orderby'])
                     ):
                         await (PatchSetting(conn)).execute(
                             self.user['id'],
@@ -201,16 +200,16 @@ class BoreholeProducerHandler(Producer):
                             request['orderby']
                         )
                     else:
-                        request['orderby'] = self.user[
-                            'setting'
-                        ]['eboreholetable']['orderby']
+                        if 'orderby' in self.user['setting']['eboreholetable']:
+                            request['orderby'] = self.user[
+                                'setting'
+                            ]['eboreholetable']['orderby']
 
                     if 'direction' in request and (
                         request['direction'] is not None
                     ) and (
-                        self.user[
-                            'setting'
-                        ]['eboreholetable']['direction'] != request['direction']
+                        ('direction' not in self.user['setting']['eboreholetable'] or
+                        self.user['setting']['eboreholetable']['direction'] != request['direction'])
                     ):
                         await (PatchSetting(conn)).execute(
                             self.user['id'],