+++ src/client/cypress/e2e/editor/location.cy.js
@@ -1,26 +1,28 @@
 import { newEditableBorehole, returnToOverview, stopBoreholeEditing } from "../helpers/testHelpers";
+import { checkRowWithText, showTableAndWaitForData } from "../helpers/dataGridHelpers";
 
 describe("Tests for 'Location' edit page.", () => {
   it("creates and deletes a borehole.", () => {
     newEditableBorehole();
 
     // enter original name
-    cy.contains("label", "Original name").next().children("input").type("SCATORPS");
+    cy.contains("label", "Original name").next().children("input").type("AAA_SCATORPS");
     cy.wait("@edit_patch");
 
     // stop editing
     stopBoreholeEditing();
     returnToOverview();
-    cy.get('[data-cy="showTableButton"]').click();
+    showTableAndWaitForData();
 
     // search the newly created borehole and delete it
     cy.get('[data-cy="borehole-table"]').within(() => {
-      cy.contains("SCATORPS").parent().find(".checkbox").click();
-      cy.contains("button", "Delete").click();
+      checkRowWithText("AAA_SCATORPS");
     });
-    cy.get(".modal button.negative").click();
+
+    cy.get('[data-cy="delete-button"]').click();
+    cy.get('.MuiButton-containedPrimary[data-cy="delete-button"]').click();
     cy.wait(["@edit_deletelist", "@edit_list"]);
-    cy.get('[data-cy="borehole-table"]').contains("SCATORPS").should("not.exist");
+    cy.get('[data-cy="borehole-table"]').contains("AAA_SCATORPS").should("not.exist");
   });
 
   it("removes error highlight of identifier fields if at least one identifier is present.", () => {