+++ src/client/cypress/e2e/filters/identifierFilter.cy.js
@@ -1,4 +1,5 @@
 import { newEditableBorehole, returnToOverview, stopBoreholeEditing } from "../helpers/testHelpers.js";
+import { checkAllVisibleRows, verifyPaginationText } from "../helpers/dataGridHelpers";
 
 describe("Tests for filtering data by identifier.", () => {
   it("can filter by identifier", () => {
@@ -20,7 +21,7 @@ describe("Tests for filtering data by identifier.", () => {
     // show all options
     cy.get('[class="ui fitted toggle checkbox"]').eq(0).children().first().check({ force: true });
 
-    cy.get("tbody").children().should("have.length", 100);
+    verifyPaginationText("1–100 of 1627");
 
     cy.get('[data-cy="domain-dropdown"]')
       .first()
@@ -29,11 +30,10 @@ describe("Tests for filtering data by identifier.", () => {
       .eq(1)
       .click({ force: true });
 
-    cy.get("tbody").children().should("have.length", 1);
-
+    verifyPaginationText("1–1 of 1");
     // click reset label
     cy.get('[data-cy="reset-filter-button"]').click();
-    cy.get("tbody").children().should("have.length", 100);
+    verifyPaginationText("1–100 of 1627");
   });
 
   it("can bulk edit boreholes while filter by identifier is set", () => {
@@ -74,11 +74,9 @@ describe("Tests for filtering data by identifier.", () => {
       .find('[role="option"]')
       .eq(1)
       .click({ force: true });
-    cy.get("tbody").children().should("have.length", 2);
-
-    cy.get('[data-cy="borehole-table"] thead .checkbox').click({ force: true });
+    verifyPaginationText("1–2 of 2");
+    checkAllVisibleRows();
     cy.contains("button", "Bulk editing").click({ force: true });
-    cy.wait("@edit_ids");
 
     // Bulk edit dialog should open.
     cy.get(".modal .toggle").should("have.length", 18);