+++ src/client/src/components/prompt/promptInterface.ts
@@ -10,13 +10,15 @@ export interface PromptAction {
 
 export interface PromptContextInterface {
   message?: string;
+  dialogContent?: ReactNode;
   actions?: PromptAction[];
   promptIsOpen: boolean;
-  showPrompt: (message: string, actions: PromptAction[]) => void;
+  showPrompt: (message: string, actions: PromptAction[], dialogContent?: ReactNode) => void;
   closePrompt: () => void;
 }
 
 export interface PromptOptions {
   message: string;
   actions: PromptAction[];
+  dialogContent: ReactNode;
 }